Remove PermissionKeyFieldEditor
