--- app/src/userpages/components/PermissionKeyField/PermissionKeyFieldEditor/permissionKeyFieldEditor.pcss
@@ -1,21 +1,0 @@
-.editor {
-  position: relative;
-  background: #F5F5F5;
-  border-radius: 4px;
-  margin: -1rem -2rem 0 -2rem;
-  padding: 2rem;
-}
-
-.buttons {
-  justify-content: flex-start;
-  padding: 0;
-  margin-top: 1.875rem;
-}
-
-.keyName {
-}
-
-.keyValue {
-  margin-top: 1.7rem;
-  max-width: calc(100% - 1rem - 130px);
-}
