--- app/src/userpages/components/PermissionKeyField/permissionKeyField.pcss
@@ -1,25 +1,0 @@
-.editor {}
-
-.root {
-  &:not(:first-child) {
-    & > .editor {
-      margin-top: 1.5rem;
-    }
-  }
-}
-
-.PermissionKeyField {}
-
-.keyFieldContainer {
-  position: relative;
-}
-
-.select {
-}
-
-.permissionHeader {
-  &:empty::after {
-    content: ' ';
-    white-space: pre;
-  }
-}
