+++ app/src/userpages/tests/components/ProfilePage/IdentityHandler/identityHandler.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { shallow, mount } from 'enzyme'
+import { mount } from 'enzyme'
 import { act } from 'react-dom/test-utils'
 
 import IdentityHandler from '../../../../components/ProfilePage/IdentityHandler'
@@ -57,7 +57,7 @@ describe('IdentityHandler', () => {
 
     describe('connect account', () => {
         it('must open dialog to connect identity', async () => {
-            const el = shallow(<IdentityHandler />)
+            const el = mount(<IdentityHandler />)
 
             await act(async () => {
                 await el.find('Button').at(0).simulate('click')
@@ -67,7 +67,7 @@ describe('IdentityHandler', () => {
         })
 
         it('must open dialog to create identity', async () => {
-            const el = shallow(<IdentityHandler />)
+            const el = mount(<IdentityHandler />)
 
             await act(async () => {
                 await el.find('Button').at(1).simulate('click')
@@ -91,7 +91,7 @@ describe('IdentityHandler', () => {
 
     describe('render', () => {
         it('should render correctly', () => {
-            const handler = shallow(<IdentityHandler />)
+            const handler = mount(<IdentityHandler />)
             const handlerSegment = handler.find('IntegrationKeyList')
             expect(handlerSegment.exists()).toBe(true)
         })
